% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catassoc.R
\name{catassoc}
\alias{catassoc}
\title{Find association between ordinal variables}
\usage{
catassoc(data, cutoff = 0.05)
}
\arguments{
\item{data}{data.table object}

\item{cutoff}{value for P cutoff, default = 0.05}
}
\description{
Finds categorical values, calculates pairwise Cramer V and filters for p < 0.05.
}
\examples{
catassoc(df)
}

