% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{concatenate without quotes (fast c replacement)}
\usage{
cc(...)
}
\arguments{
\item{character}{strings}
}
\description{
This is a utility that avoids the quotes that have to surround every element in a vector. It works only with elements that are valid names (no spaces etc).
}
\examples{
myvec = cc(one, two, three, "here you have to add quotes because there are spaces...")
}

