% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanchar.R
\name{cleanchar}
\alias{cleanchar}
\title{Clean character and factor vectors}
\usage{
cleanchar(mydf, n.factors = 25)
}
\arguments{
\item{n.factors}{how many values should the factor have?}

\item{datatable}{if you supply a data.frame it will be converted to a data.table.}
}
\value{
datatable
}
\description{
Takes character and factor vectors. Factors are first converted into characters.
Text strings are validated, converted to lower case, trimmed (trailing, leading), excessive whitespace is removed,
non printable characters are removed, encoding is set to UTF-8.
Strings "null", "NA" and " " and "  " are converted to NA.
}
\details{
factors are lumped to a number corresponding to n.factors.
NA values are converted to an explicit value ("..missing..").
All character vectors are converted to factors.
}

