% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanheader.R
\name{cleanheader}
\alias{cleanheader}
\title{Clean column names of a dataset}
\usage{
cleanheader(mydf)
}
\arguments{
\item{mydf}{data.table or data.frame. If a data.frame is supplied it will be converted to a data.table}
}
\value{
mydf data.table with cleaned column names
}
\description{
Variable names in a dataset normally need cleaning in order to:
Remove excessive whitespace (trailing, leading, double spaces)
Non alpha numeric characters. The characters are replaced by dots.
Leading and trailing dots are removed.
Encodings are set to UTF-8.
Names are set to lowercase.
}
\examples{
mydf = cleanheader(mydf)
}

