% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\title{Contrast calculate and visualise measures of inequality and concentration}
\usage{
contrast(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
Lorentz curve plot and inequality/concentration measures
}
\description{
Contrast calculate and visualise measures of inequality and concentration
}
\section{Explanation}{

For a supplied vectors of numbers, the following measures are calculated
\itemize{
    \item Lorentz curve
    \item Pen's parade
    \item Gini index
    \item Theil index
    \item Herfindahl
    \item Hoover index
}
}

\section{Details}{

The \href{https://en.wikipedia.org/wiki/Lorenz_curve}{Lorentz curve} shows the cumulative percentage of a variable against the cumulative percentage of the population.
E.g. 20 percent of the population has 10 percent of the total income.
A larger 'sag' in the line means larger inequality.

In \href{http://www.theatlantic.com/magazine/archive/2006/09/the-height-of-inequality/305089/}{Pen's parade} the values are sorted from low to high.
The plot shows the share of the population against the average for the value.
The average is indicated by a dotted line.
It can be visualised as a parade of people which lasts one hour, starting with the smallest values (e.g. incomes).

The \href{https://en.wikipedia.org/wiki/Gini_coefficient}{Gini index} summarises the inequality that is visualised in the Lorentz curve.
The index is calculated by dividing the area between the dotted line (perfect equality) and the Lorentz line by the total area under the dotted line.

The \href{https://en.wikipedia.org/wiki/Income_inequality_metrics}{Hoover index} is defined as the maximum vertical distance between the dotted line and the Lorentz curve.
It is shown as a dotted line in the Lorentz curve (percentages).
The Hoover index is the proportion of all income which would have to be redistributed to achieve a state of perfect equality.

The \href{https://en.wikipedia.org/wiki/Income_inequality_metrics}{Palma ratio} is the ratio of the 40 percent quantile and the 10 percent quantile.
The Palma ratio is thought to overcome some weaknesses of the Gini index (insensitivity for top and bottom end of the line).

The \href{https://en.wikipedia.org/wiki/Theil_index}{Theil index} is the maximum possible entropy of the data minus the observed entropy.

The \href{https://en.wikipedia.org/wiki/Herfindahl_index}{Herfindahl index} is a measure of concentration.
In its original context the Herfindahl index measures the competition between firms, with a value of larger than 0.25 considered as critical (US).
}
\examples{
data(mtcars)
contrast(mtcars$hp)
}

