% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crud.R
\docType{package}
\name{crud-package}
\alias{crud}
\alias{crud-package}
\title{Ninety percent of everything is crud, according to Sturgeon's Law.}
\description{
This package helps you to deal with ninety percent of your (raw) data.
}

