% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curry.R
\name{curry}
\alias{curry}
\title{Curry a function}
\usage{
curry(FUN, ...)
}
\arguments{
\item{function}{1, 2, i}
}
\value{
curried function
}
\description{
Curry a function. Comes from library functional.
}
\examples{
f = function(a, b) {c = a * b; return(c)}
f(2,3)
g = curry(f, b = 10)
f(2)
}

