% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup.R
\name{dedup}
\alias{dedup}
\title{Remove duplicates}
\usage{
dedup(mydf)
}
\arguments{
\item{mydf}{a data.table. Will be converted into a data.table if it isn't one.}
}
\value{
dups list of duplicates - both originals and duplicates are returned.
}
\description{
Remove duplicates
}
\examples{
find_dups = dedup(mydf)
clean_df = find_dups[[1]]
dups = find_dups[[2]]
}

