% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dempty.R
\name{dempty}
\alias{dempty}
\title{Find and remove empty columns and/or rows}
\usage{
dempty(mydf)
}
\arguments{
\item{mydf}{data.table; if it is a data.frame it will be converted into a data.table}
}
\value{
data.table without empty rows and columns
}
\description{
Find and remove empty columns and/or rows
}
\examples{
dempty(mydf)
}

