% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyplot.R
\name{dyplot}
\alias{dyplot}
\title{Timeplot using dygraph library}
\usage{
dyplot(mydf, ylabel = "count", mainlabel = "timeplot", roll = 1)
}
\arguments{
\item{mydf}{a time series, data.frame or data.table. The first column must hold the x-axis values. All other columns will be considered to be y values.}

\item{ylabel}{the label that you want to use for your y-axis}

\item{mainlabel}{the title of your plot}

\item{roll}{rolling average, default = 1 (no rolling average).}
}
\value{
dygraph time plot
}
\description{
This function is wrapper for the dygraphs library.
The purpose is to avoid code repeats for options that you always want to have anyway, such as range selectors etc.
}
\examples{
lungDeaths = cbind(mdeaths, fdeaths)
dyplot(lungDeaths, ylabel = "Number of deaths", mainlabel = "Lung deaths", roll = 7)
}

