% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_date.R
\name{excel_date}
\alias{excel_date}
\title{Unmess Excel dates}
\usage{
excel_date(mydate)
}
\arguments{
\item{character}{vector}
}
\value{
unmessed excel date vector
}
\description{
If you are importing data from Excel, you may have dates that are in a numeric format.
We can still use as.Date to import these, we simply need to know the origin date that Excel starts counting from, and provide that to as.Date.
For Excel on Windows, the origin date is December 30, 1899 for dates after 1900. (Excel’s designer thought 1900 was a leap year, but it was not.)
For Excel on Mac, the origin date is January 1, 1904.
http://www.r-bloggers.com/date-formats-in-r/
}
\examples{
myfile = structure(list(filename = c("buy", "buy", "buy", "buy", "buy"
), date = c(42248, 42248, 42248, 42248, 42248)), .Names = c("filename",
                                                            "date"), row.names = c(NA, -5L), class = c("data.table", "data.frame"
                                                            ))
myfile[, date := unmess.excel.dates(date)]
}

