% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getvalues.R
\name{getvalues}
\alias{getvalues}
\title{Get unique values from factor variables}
\usage{
getvalues(mydf)
}
\arguments{
\item{data}{data.table object}
}
\value{
vals list of unique values
}
\description{
Data table is screened for factor variables.
The unique values of each variable is obtained.
}
\examples{
checkval(data)[]
}

