% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Filter out}
\usage{
x \%nin\% table
}
\arguments{
\item{vector}{of elements to be filtered out}
}
\value{
vector of filtered elements
}
\description{
This function is the opposite of the %in% operator.
You specify the vector that you want to test on the left side, and the test values on the right side, see example.
}
\examples{
data(mtcars); setDT(mtcars)
mtcars[cyl \%in\% c(4,8), ]
mtcars[cyl \%nin\% c(4,8), ]
}

