% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modus.R
\name{modus}
\alias{modus}
\title{Modues of distribution}
\usage{
modus(x)
}
\arguments{
\item{vector}{of any type}
}
\value{
mode
}
\description{
All frequency distributions are not normally distributed and it is therefore not a good idea to use the mean.
Rather use the median (halfway) or mode (most frequent).
Source: https://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode
}
\examples{
modus(mtcars$cyl)  # most frequent type of cars: 8 cylinders
}

