% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see.R
\name{see}
\alias{see}
\title{Plot summary of numerical vectors}
\usage{
see(mydf)
}
\arguments{
\item{mydf}{a data.table object. If it isn't a data.table, it will be converted into one.}
}
\value{
plots: density, box and cumulative empirical distribution.
}
\description{
Numerical vectors in a data.table are plotted in 3 different ways: density, violin and cumulative.
The plots are then combined.
This function is inspired on the DescTools package. The motivation for this function is to avoid base plots in R as I find them clunky.
}
\examples{
data(mtcars)
see(mtcars)
}

