% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Set up environment}
\usage{
setup(trunk = trunk, create.dirs = T)
}
\arguments{
\item{none}{no parameters required}
}
\value{
none
}
\description{
Detect cores and set up cluster.
Sets locale.
Sets ggplot2 defaults.
}
\examples{
setup()
}

