% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratsamp.R
\name{stratsamp}
\alias{stratsamp}
\title{Take a stratified sample of your dataset}
\usage{
stratsamp(data, var, fraction = 0.8, split = T)
}
\arguments{
\item{var}{The variable to base your stratified sample on}

\item{fraction}{The fraction of the nrows of the dataset to sample}

\item{split}{Split dataset into a training set and a test set? Default = T}

\item{data.table}{object}
}
\value{
stratified sample of {data, var}, size nrow * fraction
}
\description{
Takes a character or factor variable and calculates the frequencies.
}
\examples{
data(mtcars)
mtcars = as.data.table(mtcars, keep.rownames = T)
stratsamp(mtcars, "cyl", 0.3, split = F)[]
}

