% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2dt.R
\name{table2dt}
\alias{table2dt}
\title{Convert table to data.table}
\usage{
table2dt(tab, label.cols = F)
}
\arguments{
\item{tab}{a table}

\item{labels.cols}{should each column be labeled with the table name?}
}
\value{
data.table a data.table object with the rownames header taken from the tab name
}
\description{
Takes a table and converts it into a matrix. The name of the table is used to name the row headers.
}
\examples{
data(mtcars)
mytab = with(mtcars, table(mpg, cyl))
table2dt(mytab)
}

