% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpackList.R
\name{unpackList}
\alias{unpackList}
\title{unpack a list}
\usage{
unpackList(mylist)
}
\arguments{
\item{mylist}{: a list object with data.tables, objects etc}
}
\value{
items from list
}
\description{
This function is from library PBSmodelling.
Atomizes a list into its constituents.
}
\examples{
l = list(a = "a", b = "b", c = "c")
unpackList(l)
}

