% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimp.R
\name{varimp}
\alias{varimp}
\title{Variable importance}
\usage{
varimp(mydf, predictor)
}
\arguments{
\item{mydf}{data.table}
}
\value{
variable importance plot
}
\description{
The Boruta algorithm is a wrapper for randomForest.
It shuffles variable values (shadows) and then compares whether the randomForest prediction improves the predictive value compared to that shadow value.
A variable importance plot is produced including the shadow variables.
}
\examples{
varimp(df, "myclass")
}

