% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray.R
\name{xray}
\alias{xray}
\title{Get overview of dataset}
\usage{
xray(mydf)
}
\arguments{
\item{data}{data.table object}
}
\value{
xray data.table with the fields {variable name, class, blanks, infinites, NaNs, zeros, means, medians, most frequent values, fraction most frequent values}.
}
\description{
List all characteristics of data, split out by class.
For numeric values: mean, median, infinites, NaNs, negatives, zeros
for factors: levels and unused levels;
all variables: class, blanks.
}
\examples{
data(ggplot2::msleep)
msleep = as.data.table(msleep, keep.rownames = T)
xray(msleep)[]
}

